Agregador_APIS/api_clients/requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
python-dotenv>=1.0.0